r591: patch for a wrong casting config.offset_fixed_value for the textbox - my mistake

+++ plugins/seltempavg/seltempavg.h
@@ -41,7 +41,7 @@ public:
 	int paranoid;
 	int nosubtract;
 	int offset_restartmarker_keyframe;
-	int offset_fixed_value;
+	int64_t offset_fixed_value;
 	float gain;
 
 };
